Inpaint-SDXL/requirements.txt
# SDXL Inpainting API Requirements
# Install with: pip install -r requirements.txt

# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.21.0
transformers>=4.30.0
accelerate>=0.20.0

# API server
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
python-multipart>=0.0.6

# Image processing
Pillow>=9.5.0
numpy>=1.24.0

# Optional: for better performance
xformers>=0.0.20  # Memory efficient attention (optional)
